完成发布商品以及展示

+++ campus-api/src/main/java/org/kuro/campus/model/page/PageResult.java
@@ -8,22 +8,22 @@ import java.util.List;
  */
 public class PageResult<T> {
 
-    private Long total;
+    private Integer total;
     private List<T> items;
 
     public PageResult() {
     }
 
-    public PageResult(Long total, List<T> items) {
+    public PageResult(Integer total, List<T> items) {
         this.total = total;
         this.items = items;
     }
 
-    public Long getTotal() {
+    public Integer getTotal() {
         return total;
     }
 
-    public void setTotal(Long total) {
+    public void setTotal(Integer total) {
         this.total = total;
     }
